Get rid of SimXXXVariablePhi classes. Update VariableRecovery analysis.
